requirements.txt
fastapi==0.85.1
pandas==1.4.4
psycopg2==2.9.5
psycopg2_binary==2.9.5
uvicorn==0.19.0
PyYAML==6.0